+++ CameraPan/HarmonyPatches/EventPatches.cs
@@ -10,5 +10,5 @@ internal static class EventPatches
 {
     [HarmonyPatch(typeof(Game1), nameof(Game1.eventFinished))]
     private static void PostfixEventEnd()
-        => ModEntry.snapOnNextTick.Value = true;
+        => ModEntry.SnapOnNextTick = true;
 }