+++ GrowableBushes/docs/changelog.md
@@ -4,6 +4,9 @@
 ### TODO
    - Consider making the bushes their own tab in the furniture catalogue?
 
+### Version 0.1.1
+* Added option to allow bushes in greenhouses to bloom all year.
+
 ### Version 0.1.0
 
 Initial Upload.